requirements.txt
# Web
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6

# Data / config
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0

# LLM（Ollama / OpenAI 兼容云端共用 openai 客户端）
openai>=1.40.0

# Agent：Function Calling（LangChain）；LLM_MODE=off 时不加载
langchain-core>=0.3.0
langchain-openai>=0.2.0

# Dev
pytest>=7.4.0
black>=23.0.0
